BE-0: move ProfileType to separate common module
