+++ sources/common/common_profile/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    id("photo.world.kotlin.lib")
+}
