+++ sources/domain/domain_profile/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 
 dependencies {
     implementation(project(":common_errors"))
+    api(project(":common_profile"))
 
     testImplementation(libs.bundles.kotest)
 }
