+++ sources/domain/domain_profile/src/main/kotlin/photo/world/domain/profile/ext/ListExt.kt
@@ -1,5 +1,6 @@
 package photo.world.domain.profile.ext
 
+import photo.world.common.profile.ProfileType
 import photo.world.domain.profile.entity.Album
 import photo.world.domain.profile.entity.profile.*
