+++ sources/domain/domain_profile/src/main/kotlin/photo/world/domain/profile/service/AlbumService.kt
@@ -1,7 +1,7 @@
 package photo.world.domain.profile.service
 
+import photo.world.common.profile.ProfileType
 import photo.world.domain.profile.entity.Album
-import photo.world.domain.profile.entity.profile.ProfileType
 
 interface AlbumService {
