+++ sources/domain/domain_profile/src/main/kotlin/photo/world/domain/profile/service/ProfileService.kt
@@ -1,5 +1,6 @@
 package photo.world.domain.profile.service
 
+import photo.world.common.profile.ProfileType
 import photo.world.domain.profile.entity.profile.*
 import photo.world.domain.profile.service.data.ProfileData
