+++ sources/domain/domain_profile/src/main/kotlin/photo/world/domain/profile/service/impl/DomainAlbumService.kt
@@ -1,8 +1,8 @@
 package photo.world.domain.profile.service.impl
 
+import photo.world.common.profile.ProfileType
 import photo.world.domain.profile.entity.Album
 import photo.world.domain.profile.entity.profile.Profile
-import photo.world.domain.profile.entity.profile.ProfileType
 import photo.world.domain.profile.repository.AccountRepository
 import photo.world.domain.profile.repository.ProfileRepository
 import photo.world.domain.profile.service.AlbumService
