+++ sources/domain/domain_profile/src/main/kotlin/photo/world/domain/profile/service/impl/DomainProfileContentService.kt
@@ -1,9 +1,9 @@
 package photo.world.domain.profile.service.impl
 
+import photo.world.common.profile.ProfileType
 import photo.world.domain.profile.entity.Service
 import photo.world.domain.profile.entity.profile.ModelProfile
 import photo.world.domain.profile.entity.profile.Profile
-import photo.world.domain.profile.entity.profile.ProfileType
 import photo.world.domain.profile.repository.AccountRepository
 import photo.world.domain.profile.repository.ProfileRepository
 import photo.world.domain.profile.service.ProfileContentService
