+++ sources/infrastructure/data_user/src/main/kotlin/photo/world/data/user/repository/postgresql/profile/PostgresqlProfileRepository.kt
@@ -1,6 +1,7 @@
 package photo.world.data.user.repository.postgresql.profile
 
 import org.springframework.stereotype.Component
+import photo.world.common.profile.ProfileType
 import photo.world.data.user.entity.profile.DataProfile
 import photo.world.data.user.entity.profile.ProfileServiceRelationship
 import photo.world.data.user.entity.user.BaseUser
@@ -15,7 +16,6 @@ import photo.world.domain.errors.NotFoundEntityException
 import photo.world.domain.errors.notFound
 import photo.world.domain.profile.entity.profile.LiteProfile
 import photo.world.domain.profile.entity.profile.Profile
-import photo.world.domain.profile.entity.profile.ProfileType
 import photo.world.domain.profile.repository.ProfileRepository
 
 @Component
