+++ sources/infrastructure/data_user/src/main/kotlin/photo/world/data/user/repository/postgresql/profile/PostgresqlServiceRepository.kt
@@ -1,8 +1,8 @@
 package photo.world.data.user.repository.postgresql.profile
 
 import org.springframework.stereotype.Component
+import photo.world.common.profile.ProfileType
 import photo.world.data.user.repository.spring.profile.SpringDataPostgresqlServiceRepository
-import photo.world.domain.profile.entity.profile.ProfileType
 import photo.world.domain.profile.repository.ServiceRepository
 
 @Component
