+++ sources/infrastructure/data_user/src/main/kotlin/photo/world/data/user/repository/postgresql/profile/PostgresqlTagRepository.kt
@@ -1,9 +1,9 @@
 package photo.world.data.user.repository.postgresql.profile
 
 import org.springframework.stereotype.Component
+import photo.world.common.profile.ProfileType
 import photo.world.data.user.repository.spring.profile.SpringDataPostgresqlTagRepository
 import photo.world.domain.profile.entity.Tag
-import photo.world.domain.profile.entity.profile.ProfileType
 import photo.world.domain.profile.repository.TagRepository
 
 @Component
